Pquery - use asyncio by default, offer alternatives
... via 'extra' package. The differences are very small though
